# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AppConfig {
  /** When true, a verified phone login is required; when false the app may offer demo sign-in */
  requireAuth: boolean;
}

export interface RegisterShopRequest {
  /**
   * @minLength 2
   * @maxLength 80
   */
  name: string;
  /**
   * @minLength 2
   * @maxLength 80
   */
  ownerName: string;
  /**
   * @minLength 5
   * @maxLength 200
   */
  address: string;
  /**
   * @minimum -90
   * @maximum 90
   */
  lat: number;
  /**
   * @minimum -180
   * @maximum 180
   */
  lng: number;
  /** @pattern ^([01]\d|2[0-3]):[0-5]\d$ */
  openTime: string;
  /** @pattern ^([01]\d|2[0-3]):[0-5]\d$ */
  closeTime: string;
  /**
   * @minItems 1
   * @maxItems 12
   */
  categories: string[];
}

export interface ProductInput {
  /**
   * @minLength 1
   * @maxLength 120
   */
  name: string;
  /** @exclusiveMinimum 0 */
  price: number;
  /**
   * @minLength 1
   * @maxLength 20
   */
  unit: string;
  /**
   * @minLength 1
   * @maxLength 40
   */
  category: string;
  /** @minimum 0 */
  stock: number;
  /** @maxLength 1000 */
  description?: string;
  isWeightBased?: boolean;
  isActive?: boolean;
}

export interface ProductUpdate {
  /**
   * @minLength 1
   * @maxLength 120
   */
  name?: string;
  /** @exclusiveMinimum 0 */
  price?: number;
  /**
   * @minLength 1
   * @maxLength 20
   */
  unit?: string;
  /**
   * @minLength 1
   * @maxLength 40
   */
  category?: string;
  /** @minimum 0 */
  stock?: number;
  /** @maxLength 1000 */
  description?: string;
  isWeightBased?: boolean;
  isActive?: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface ApiErrorResponse {
  message: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  accepted: "accepted",
  packed: "packed",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  rejected: "rejected",
} as const;

export interface OrderItem {
  productId: string;
  name: string;
  price: number;
  unit: string;
  quantity: number;
  selectedWeight?: string;
}

export type OrderMode = (typeof OrderMode)[keyof typeof OrderMode];

export const OrderMode = {
  pickup: "pickup",
  delivery: "delivery",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cod: "cod",
  upi: "upi",
} as const;

export interface Order {
  id: string;
  shopId: string;
  shopName: string;
  customerPhone: string;
  items: OrderItem[];
  total: number;
  deliveryFee: number;
  status: OrderStatus;
  mode: OrderMode;
  address?: string | null;
  paymentMethod: OrderPaymentMethod;
  placedAt: string;
}

export interface UpdateOrderStatusRequest {
  status: OrderStatus;
}

export type PlaceOrderRequestMode =
  (typeof PlaceOrderRequestMode)[keyof typeof PlaceOrderRequestMode];

export const PlaceOrderRequestMode = {
  pickup: "pickup",
  delivery: "delivery",
} as const;

export type PlaceOrderRequestPaymentMethod =
  (typeof PlaceOrderRequestPaymentMethod)[keyof typeof PlaceOrderRequestPaymentMethod];

export const PlaceOrderRequestPaymentMethod = {
  cod: "cod",
  upi: "upi",
} as const;

export type PlaceOrderRequestItemsItem = {
  productId: string;
  /** @minimum 1 */
  quantity: number;
  selectedWeight?: string;
};

export interface PlaceOrderRequest {
  shopId: string;
  customerPhone: string;
  mode: PlaceOrderRequestMode;
  address?: string;
  paymentMethod: PlaceOrderRequestPaymentMethod;
  /** @minItems 1 */
  items: PlaceOrderRequestItemsItem[];
}

export interface Shop {
  id: string;
  name: string;
  address: string;
  lat: number;
  lng: number;
  rating: number;
  openTime: string;
  closeTime: string;
  isOpen: boolean;
  categories: string[];
  image?: string | null;
  ownerName?: string | null;
}

export interface Product {
  id: string;
  shopId: string;
  shopName: string;
  name: string;
  price: number;
  unit: string;
  category: string;
  stock: number;
  description?: string | null;
  image?: string | null;
  isWeightBased: boolean;
  isActive: boolean;
}

export type ListOrdersParams = {
  phone: string;
};