# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: shops
    description: Shop discovery
  - name: products
    description: Product catalog
  - name: orders
    description: Order placement and lifecycle
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /config:
    get:
      operationId: getAppConfig
      tags: [health]
      summary: Public app configuration
      description: Settings the app needs before sign-in (e.g. whether login is enforced)
      responses:
        "200":
          description: Configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppConfig"
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List shops
      description: Returns all shops
      responses:
        "200":
          description: Shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
    post:
      operationId: registerShop
      tags: [shops]
      summary: Register a shop
      description: Creates a shop owned by the calling shopkeeper (one shop per owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterShopRequest"
      responses:
        "201":
          description: Shop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Caller not identified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Caller already owns a shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /me/shop:
    get:
      operationId: getMyShop
      tags: [shops]
      summary: Get the caller's shop
      description: Returns the shop owned by the calling shopkeeper
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "404":
          description: Caller owns no shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /shops/{shopId}/inventory:
    get:
      operationId: listShopInventory
      tags: [products]
      summary: List a shop's full inventory
      description: All products including inactive ones; shop owner only
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "403":
          description: Not the shop's owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /shops/{shopId}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get a shop
      description: Returns a single shop by id
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "404":
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /shops/{shopId}/products:
    get:
      operationId: listShopProducts
      tags: [products]
      summary: List a shop's products
      description: Returns all active products of a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "404":
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Add a product to a shop
      description: Shop owner only
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Not the shop's owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /orders:
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order
      description: Creates an order; prices are computed server-side from the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Shop or product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    get:
      operationId: listOrders
      tags: [orders]
      summary: List a customer's orders
      description: Returns orders for a customer phone, newest first
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update an order's status
      description: Shopkeeper action; only forward transitions are allowed
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /shops/{shopId}/orders:
    get:
      operationId: listShopOrders
      tags: [orders]
      summary: List a shop's orders
      description: Returns all orders of a shop, newest first
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "404":
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      description: Returns a single product by id
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      description: Shop owner only; send only the fields to change
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Not the shop's owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      description: Shop owner only
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "403":
          description: Not the shop's owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
components:
  schemas:
    AppConfig:
      type: object
      properties:
        requireAuth:
          type: boolean
          description: When true, a verified phone login is required; when false the app may offer demo sign-in
      required:
        - requireAuth
    RegisterShopRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        ownerName:
          type: string
          minLength: 2
          maxLength: 80
        address:
          type: string
          minLength: 5
          maxLength: 200
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        openTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        closeTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        categories:
          type: array
          minItems: 1
          maxItems: 12
          items:
            type: string
      required:
        - name
        - ownerName
        - address
        - lat
        - lng
        - openTime
        - closeTime
        - categories
    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        price:
          type: number
          exclusiveMinimum: 0
        unit:
          type: string
          minLength: 1
          maxLength: 20
        category:
          type: string
          minLength: 1
          maxLength: 40
        stock:
          type: integer
          minimum: 0
        description:
          type: string
          maxLength: 1000
        isWeightBased:
          type: boolean
        isActive:
          type: boolean
      required:
        - name
        - price
        - unit
        - category
        - stock
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        price:
          type: number
          exclusiveMinimum: 0
        unit:
          type: string
          minLength: 1
          maxLength: 20
        category:
          type: string
          minLength: 1
          maxLength: 40
        stock:
          type: integer
          minimum: 0
        description:
          type: string
          maxLength: 1000
        isWeightBased:
          type: boolean
        isActive:
          type: boolean
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    OrderStatus:
      type: string
      enum:
        - pending
        - accepted
        - packed
        - out_for_delivery
        - delivered
        - rejected
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        price:
          type: number
        unit:
          type: string
        quantity:
          type: integer
        selectedWeight:
          type: string
      required:
        - productId
        - name
        - price
        - unit
        - quantity
    Order:
      type: object
      properties:
        id:
          type: string
        shopId:
          type: string
        shopName:
          type: string
        customerPhone:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        deliveryFee:
          type: number
        status:
          $ref: "#/components/schemas/OrderStatus"
        mode:
          type: string
          enum: [pickup, delivery]
        address:
          type: string
          nullable: true
        paymentMethod:
          type: string
          enum: [cod, upi]
        placedAt:
          type: string
          format: date-time
      required:
        - id
        - shopId
        - shopName
        - customerPhone
        - items
        - total
        - deliveryFee
        - status
        - mode
        - paymentMethod
        - placedAt
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
      required:
        - status
    PlaceOrderRequest:
      type: object
      properties:
        shopId:
          type: string
        customerPhone:
          type: string
        mode:
          type: string
          enum: [pickup, delivery]
        address:
          type: string
        paymentMethod:
          type: string
          enum: [cod, upi]
        items:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
                minimum: 1
              selectedWeight:
                type: string
            required:
              - productId
              - quantity
      required:
        - shopId
        - customerPhone
        - mode
        - paymentMethod
        - items
    Shop:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        rating:
          type: number
        openTime:
          type: string
        closeTime:
          type: string
        isOpen:
          type: boolean
        categories:
          type: array
          items:
            type: string
        image:
          type: string
          nullable: true
        ownerName:
          type: string
          nullable: true
      required:
        - id
        - name
        - address
        - lat
        - lng
        - rating
        - openTime
        - closeTime
        - isOpen
        - categories
    Product:
      type: object
      properties:
        id:
          type: string
        shopId:
          type: string
        shopName:
          type: string
        name:
          type: string
        price:
          type: number
        unit:
          type: string
        category:
          type: string
        stock:
          type: integer
        description:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        isWeightBased:
          type: boolean
        isActive:
          type: boolean
      required:
        - id
        - shopId
        - shopName
        - name
        - price
        - unit
        - category
        - stock
        - isWeightBased
        - isActive