# lib/api-zod/src/generated/types/appConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppConfig {
  /** When true, a verified phone login is required; when false the app may offer demo sign-in */
  requireAuth: boolean;
}